+++ pushy/src/main/java/com/eatthepath/pushy/apns/PushType.java
@@ -23,6 +23,11 @@
 package com.eatthepath.pushy.apns;
 
 import com.eatthepath.pushy.apns.auth.ApnsSigningKey;
+import com.eatthepath.pushy.apns.util.ApnsPayloadBuilder;
+import com.eatthepath.pushy.apns.util.LiveActivityEvent;
+
+import java.time.Instant;
+import java.util.Map;
 
 /**
  * An enumeration of push notification display types. Note that push notification display types are required in iOS 13
@@ -142,17 +147,15 @@ public enum PushType {
     LOCATION("location"),
 
     /**
-     * <p>Indicates that a push notification is intended to update a running Live Activity. According to Apple's
-     * documentation:</p>
-     *
-     * <blockquote><p>Set the apns-topic header field of the request you sent to APNs using the following format:
-     * {@code <your bundleID>.push-type.liveactivity.}</p>
-     *
-     * <p>To update a Live Activity, set the value for the payload’s event key to update. To end a Live Activity,
-     * set it to end. If you end a Live Activity, include the final content state to make sure the
-     * Live Activity displays the latest data after it ends.</p></blockquote>
+     * Indicates that a push notification is intended to update a running Live Activity. Note that Live Activity updates
+     * must be sent to a specific topic ({@code [base bundle ID].push-type.liveactivity}) and include specific payload
+     * keys (see {@link ApnsPayloadBuilder#setContentState(Map)}, {@link ApnsPayloadBuilder#setEvent(LiveActivityEvent)},
+     * {@link ApnsPayloadBuilder#setDismissalDate(Instant)}, and {@link ApnsPayloadBuilder#setTimestamp(Instant)}).
      *
      * @since 0.15.2
+     *
+     * @see <a href="https://developer.apple.com/documentation/activitykit/update-and-end-your-live-activity-with-remote-push-notifications">
+     *     Updating and ending your Live Activity with remote push notifications</a>
      */
     LIVE_ACTIVITY("liveactivity");
